refactor: updated component folder names to better match new standard
