--- src/components/utils/index.tsx
@@ -1,14 +1,0 @@
-/* ================================================================
-Description: Index file for utitlity components
-Author: Noah Huesman
-
-Creation Date: 01/13/2025
-Modification History:
-#1 (01/13/2025) - Initial creation - Noah Huesman
-================================================================ */
-
-// ========================================
-// ERROR 404
-// ========================================
-
-export * from "./error-404"
